+++ framework/packages/abstract-std/src/native/ibc/ibc_client.rs
@@ -87,6 +87,9 @@ pub enum ExecuteMsg {
         /// host chain to be executed on
         /// Example: "osmosis"
         host_chain: TruncatedChainId,
+        /// Address of the token receiver on host chain
+        /// Defaults to address of the remote account
+        receiver: Option<String>,
         memo: Option<String>,
     },
     /// Only callable by Account