+++ framework/docs/src/releases/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Added
 
-- Abstract Account authentication when building account with `xion` feature.
+- Added `PfmMemoBuilder` API for building middleware forwarding memo
+- Added `HookMemoBuilder` API for building wasm ibc hook memo
 - `execute_with_funds` to Executor to attach funds to execution.
 - `stargate` feature for abstract-app, abstract-standalone and abstract-adapter packages.
 - New module type: `Service`, behaves the same as Native, but can be registered by any namespace.